add ticket checker to org model

+++ internal/apis/zendesk/search.go
@@ -11,7 +11,7 @@ import (
 
 type SearchQuery struct {
 	Tags                  []string
-	TicketsOrganizationId int
+	TicketsOrganizationId int64
 	TicketCreatedAfter    time.Time
 	TicketCreatedBefore   time.Time
 }
